+++ src/main/java/org/jruby/rack/RackEnvironment.java
@@ -156,4 +156,10 @@ public interface RackEnvironment {
      * @return the remote user
      */
     String getRemoteUser();
+
+    /**
+     * @see jakarta.servlet.http.HttpServletRequest#getProtocol()
+     * @return a String containing the name of the scheme used to make this request
+     */
+    String getProtocol();
 }