+++ src/spec/ruby/rack/handler/servlet_spec.rb
@@ -44,10 +44,10 @@ describe Rack::Handler::Servlet do
 
     it "creates a hash with the Rack variables in it" do
       hash = servlet.create_env(@servlet_env)
-      expect(hash['rack.version']).to eq Rack::VERSION
-      expect(hash['rack.multithread']).to eq true
-      expect(hash['rack.multiprocess']).to eq false
-      expect(hash['rack.run_once']).to eq false
+      expect(hash['rack.version']).to eq Rack.release < '3' ? Rack::VERSION : nil
+      expect(hash['rack.multithread']).to eq Rack.release < '3' ? true : nil
+      expect(hash['rack.multiprocess']).to eq Rack.release < '3' ? false : nil
+      expect(hash['rack.run_once']).to eq Rack.release < '3' ? false : nil
       expect(hash['rack.hijack?']).to eq false
     end
 
@@ -69,6 +69,7 @@ describe Rack::Handler::Servlet do
         "SERVER_NAME" => "override",
         "SERVER_PORT" => 8080,
         "SERVER_SOFTWARE" => "servy",
+        "SERVER_PROTOCOL" => "HTTP/2.0",
         "REMOTE_HOST" => "override",
         "REMOTE_ADDR" => "192.168.0.1",
         "REMOTE_USER" => "override"
@@ -84,6 +85,7 @@ describe Rack::Handler::Servlet do
       expect(env["SERVER_NAME"]).to eq "override"
       expect(env["SERVER_PORT"]).to eq "8080"
       expect(env["SERVER_SOFTWARE"]).to eq "servy"
+      expect(env["SERVER_PROTOCOL"]).to eq "HTTP/2.0"
       expect(env["REMOTE_HOST"]).to eq "override"
       expect(env["REMOTE_ADDR"]).to eq "192.168.0.1"
       expect(env["REMOTE_USER"]).to eq "override"
@@ -164,6 +166,7 @@ describe Rack::Handler::Servlet do
       @servlet_request.setQueryString('hello=there')
       @servlet_request.setServerName('serverhost')
       @servlet_request.setServerPort(80)
+      @servlet_request.setProtocol('HTTP/1.1')
       @servlet_request.setRemoteAddr('127.0.0.1')
       @servlet_request.setRemoteHost('localhost')
       @servlet_request.setRemoteUser('admin')
@@ -177,6 +180,7 @@ describe Rack::Handler::Servlet do
       expect(env["QUERY_STRING"]).to eq "hello=there"
       expect(env["SERVER_NAME"]).to eq "serverhost"
       expect(env["SERVER_PORT"]).to eq "80"
+      expect(env["SERVER_PROTOCOL"]).to eq "HTTP/1.1"
       expect(env["REMOTE_HOST"]).to eq "localhost"
       expect(env["REMOTE_ADDR"]).to eq "127.0.0.1"
       expect(env["REMOTE_USER"]).to eq "admin"
@@ -194,6 +198,7 @@ describe Rack::Handler::Servlet do
       @servlet_request.setQueryString('hello=there')
       @servlet_request.setServerName('serverhost')
       @servlet_request.setServerPort(80)
+      @servlet_request.setProtocol('HTTP/1.1')
       @servlet_request.setRemoteAddr('127.0.0.1')
       @servlet_request.setRemoteHost('localhost')
       @servlet_request.setRemoteUser('admin')
@@ -205,7 +210,7 @@ describe Rack::Handler::Servlet do
       end
 
       env = servlet.create_env @servlet_env
-      expect(env["rack.version"]).to eq Rack::VERSION
+      expect(env["rack.version"]).to eq Rack.release < '3' ? Rack::VERSION : nil
       expect(env["CONTENT_TYPE"]).to eq "text/html"
       expect(env["HTTP_HOST"]).to eq "serverhost"
       expect(env["HTTP_ACCEPT"]).to eq "text/*"
@@ -216,6 +221,7 @@ describe Rack::Handler::Servlet do
       expect(env["QUERY_STRING"]).to eq "hello=there"
       expect(env["SERVER_NAME"]).to eq "serverhost"
       expect(env["SERVER_PORT"]).to eq "80"
+      expect(env["SERVER_PROTOCOL"]).to eq "HTTP/1.1"
       expect(env["REMOTE_HOST"]).to eq "localhost"
       expect(env["REMOTE_ADDR"]).to eq "127.0.0.1"
       expect(env["REMOTE_USER"]).to eq "admin"
@@ -404,6 +410,17 @@ describe Rack::Handler::Servlet do
       expect { env.fetch('attr4') }.to raise_error # KeyError
     end
 
+    it "joins the values of repeated request headers" do
+      @servlet_request.addHeader "X-Forwarded-For", "10.0.0.1"
+      @servlet_request.addHeader "X-Forwarded-For", "10.0.0.2"
+      @servlet_request.addHeader "Cookie", "foo=1"
+      @servlet_request.addHeader "Cookie", "bar=2"
+
+      env = servlet.create_env(@servlet_env)
+      expect(env['HTTP_X_FORWARDED_FOR']).to eq "10.0.0.1, 10.0.0.2"
+      expect(env['HTTP_COOKIE']).to eq "foo=1; bar=2" # RFC 6265 cookie separator
+    end
+
   end
 
   shared_examples "(eager)rack-env" do
@@ -455,6 +472,7 @@ describe Rack::Handler::Servlet do
       expect(env.keys).to include('QUERY_STRING')
       expect(env.keys).to include('SERVER_NAME')
       expect(env.keys).to include('SERVER_PORT')
+      expect(env.keys).to include('SERVER_PROTOCOL')
       expect(env.keys).to include('REMOTE_HOST')
       expect(env.keys).to include('REMOTE_ADDR')
       expect(env.keys).to include('REMOTE_USER')
@@ -462,12 +480,16 @@ describe Rack::Handler::Servlet do
         expect(env.keys).to include(key)
       end
 
-      expect(env.keys).to include('rack.version')
+      if Rack.release < '3'
+        expect(env.keys).to include('rack.version')
+        expect(env.keys).to include('rack.multithread')
+        expect(env.keys).to include('rack.multiprocess')
+        expect(env.keys).to include('rack.run_once')
+      end
+
       expect(env.keys).to include('rack.input')
       expect(env.keys).to include('rack.errors')
       expect(env.keys).to include('rack.url_scheme')
-      expect(env.keys).to include('rack.multithread')
-      expect(env.keys).to include('rack.run_once')
       expect(env.keys).to include('java.servlet_context')
       expect(env.keys).to include('java.servlet_request')
       expect(env.keys).to include('java.servlet_response')
@@ -494,11 +516,15 @@ describe Rack::Handler::Servlet do
       expect { env['OTHER_METHOD'] }.to_not raise_error
       expect(env['OTHER_METHOD']).to be nil
 
-      expect { env['rack.version'] }.to_not raise_error
+      if Rack.release < '3'
+        expect { env['rack.version'] }.to_not raise_error
+        expect { env['rack.multithread'] }.to_not raise_error
+        expect { env['rack.multiprocess'] }.to_not raise_error
+        expect { env['rack.run_once'] }.to_not raise_error
+      end
+
       expect { env['rack.input'] }.to_not raise_error
       expect { env['rack.errors'] }.to_not raise_error
-      expect { env['rack.run_once'] }.to_not raise_error
-      expect { env['rack.multithread'] }.to_not raise_error
       expect { env['java.servlet_context'] }.to_not raise_error
       expect { env['java.servlet_request'] }.to_not raise_error
       expect { env['java.servlet_response'] }.to_not raise_error
@@ -624,17 +650,21 @@ describe Rack::Handler::Servlet do
         expect(env['SCRIPT_NAME']).to eql '/main'
         expect(env['SERVER_NAME']).to eql 'serverhost'
         expect(env['SERVER_PORT']).to eql '80'
+        expect(env['SERVER_PROTOCOL']).to eql 'HTTP/1.1'
         expect(env['OTHER_METHOD']).to be nil
         Rack::Handler::Servlet::DefaultEnv::VARIABLES.each do |key|
           expect(env[key]).to_not be(nil), "key: #{key.inspect} nil"
         end
 
         expect(env['rack.url_scheme']).to_not be nil
-        expect(env['rack.version']).to_not be nil
         expect(env['jruby.rack.version']).to_not be nil
 
-        expect(env['rack.run_once']).to be false
-        expect(env['rack.multithread']).to be true
+        if Rack.release < '3'
+          expect(env['rack.version']).to_not be nil
+          expect(env['rack.multithread']).to be true
+          expect(env['rack.multiprocess']).to be false
+          expect(env['rack.run_once']).to be false
+        end
 
         expect(env['rack.whatever']).to be nil
 
@@ -765,6 +795,7 @@ describe Rack::Handler::Servlet do
       expect(env.keys).to include('QUERY_STRING')
       expect(env.keys).to include('SERVER_NAME')
       expect(env.keys).to include('SERVER_PORT')
+      expect(env.keys).to include('SERVER_PROTOCOL')
       expect(env.keys).to include('REMOTE_HOST')
       expect(env.keys).to include('REMOTE_ADDR')
       expect(env.keys).to include('REMOTE_USER')
@@ -772,12 +803,16 @@ describe Rack::Handler::Servlet do
         expect(env.keys).to include(key)
       end
 
-      expect(env.keys).to include('rack.version')
+      if Rack.release < '3'
+        expect(env.keys).to include('rack.version')
+        expect(env.keys).to include('rack.multithread')
+        expect(env.keys).to include('rack.multiprocess')
+        expect(env.keys).to include('rack.run_once')
+      end
+
       expect(env.keys).to include('rack.input')
       expect(env.keys).to include('rack.errors')
       expect(env.keys).to include('rack.url_scheme')
-      expect(env.keys).to include('rack.multithread')
-      expect(env.keys).to include('rack.run_once')
       expect(env.keys).to include('java.servlet_context')
       expect(env.keys).to include('java.servlet_request')
       expect(env.keys).to include('java.servlet_response')
@@ -921,11 +956,46 @@ describe Rack::Handler::Servlet do
                                           "name" => ["Ferko Suska", "Jozko Hruska"], "formula" => "a + b == 42%!"
                                         })
 
+      if rack_request.respond_to?(:form_pairs) # Rack 3.2+
+        # POST name/value pairs, preserving duplicate (raw, un-nested) names,
+        # available even though the servlet input stream was already consumed
+        expect(rack_request.form_pairs).to match_array([
+          [ 'name[]', 'Ferko Suska' ], [ 'name[]', 'Jozko Hruska' ],
+          [ 'age', '30' ], [ 'formula', 'a + b == 42%!' ]
+        ])
+      end
+
       expect(rack_request.query_string).to eq 'foo=bad&foo=bar&bar=huu&age=33'
       expect(rack_request.request_method).to eq 'POST'
       expect(rack_request.path_info).to eq '/path'
       expect(rack_request.script_name).to eq '/home' # context path
       expect(rack_request.content_length).to eq content.size.to_s
+    end
+
+    it "has correct hash-in-array params when request input has been read" do
+      # the Rails nested-attributes form shape, e.g. fields_for with an array;
+      # used to raise TypeError with the (contorted) Rack 3 params algorithm
+      skip "Rack 2.x parameter mapping does not support hash-in-array params" if Rack.release < '3'
+
+      content = 'book%5Bchapters%5D%5B%5D%5Btitle%5D=first&book%5Bchapters%5D%5B%5D%5Btitle%5D=second'
+      servlet_request.setContent content.to_java_bytes
+      servlet_request.addHeader('CONTENT-TYPE', 'application/x-www-form-urlencoded')
+      servlet_request.setMethod 'POST'
+      servlet_request.setContextPath '/home'
+      servlet_request.setPathInfo '/path'
+      servlet_request.setRequestURI '/home/path'
+      # NOTE: assume input stream read but getParameter methods work correctly :
+      read_input_stream servlet_request.getInputStream
+      servlet_request.addParameter('book[chapters][][title]', 'first')
+      servlet_request.addParameter('book[chapters][][title]', 'second')
+
+      env = servlet.create_env(servlet_env)
+      rack_request = Rack::Request.new(env)
+
+      expect(rack_request.GET).to eq({})
+      expect(rack_request.POST).to eq({
+        'book' => { 'chapters' => [ { 'title' => 'first' }, { 'title' => 'second' } ] }
+      })
     end
 
     it "handles null values in parameter-map (Jetty)" do